Always deal with arrays of element at the middleware level
This makes the API more uniform.
Previously, elements could be either an array or a single element which forces
middleware authors to handle both cases.

For the users, when calling the test() method, helpers.elements will always contain arrays.
However, they can still benefit from the shortcut syntax test({myElement}).
In that case, elements will be 'unnaraid', which means that that they can either get an array
of elements or a single element depending on their component layout.